+++ test/case/cli/system-output/run/bridge_-j_link_show_dev_e0
