+++ test/case/cli/system-output/run/bridge_-j_vlan
