+++ test/case/cli/system-output/run/bridge_-j_vlan_show_dev_e2
