+++ test/case/cli/system-output/run/ethtool_--json_-S_br0_--all-groups
