+++ test/case/cli/system-output/run/ethtool_--json_-S_dsa0_--all-groups
