+++ test/case/cli/system-output/run/ethtool_--json_-S_e0_--all-groups
