+++ test/case/cli/system-output/run/ethtool_--json_-S_e1_--all-groups
