+++ test/case/cli/system-output/run/ethtool_--json_-S_e3_--all-groups
