+++ test/case/cli/system-output/run/ethtool_--json_-S_e4_--all-groups
