+++ test/case/cli/system-output/run/ethtool_br0
