+++ test/case/cli/system-output/run/ethtool_dsa0
