+++ test/case/cli/system-output/run/ethtool_e0
