+++ test/case/cli/system-output/run/ip_-j_addr_show
