+++ test/case/cli/system-output/run/ip_-j_netns_list
