+++ test/case/cli/system-output/run/ip_-s_-d_-j_link_show
