+++ test/case/cli/system-output/run/ip_netns_exec_57ff63cb_ip_-j_addr_show
