+++ test/case/cli/system-output/run/ip_netns_exec_57ff63cb_ip_-s_-d_-j_link_show
