+++ test/case/cli/system-output/run/ip_netns_exec_db5ad90e_ip_-j_addr_show
