+++ test/case/cli/system-output/run/ip_netns_exec_db5ad90e_ip_-s_-d_-j_link_show
