+++ test/case/cli/system-output/run/mctl_-p_show_igmp_json
