+++ test/case/cli/system-output/run/vtysh_-c_show-ip-route-json
