+++ test/case/cli/system-output/run/vtysh_-c_show-ipv6-route-json
